requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.24
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.2
django-toolbelt==0.0.1
gunicorn==19.6.0
kombu==3.0.37
Pillow==3.4.1
psycopg2==2.7.1
python-dateutil==2.5.3
pytz==2016.7
reportlab==3.3.0
six==1.10.0
static3==0.7.0
whitenoise==3.2.2
